requirements.txt
pymem
dnspython~=2.8.0
requests~=2.32.5
matplotlib~=3.10.7
nbtlib~=2.0.4
pillow~=12.0.0
icmplib~=3.0.4
psutil~=7.1.1
chardet~=5.2.0
openai~=2.6.0
mcstatus~=12.0.6
mcrcon~=0.7.0